// school-udemy/src/generated-sources/api/model/calificacion.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * WebApi
 * OpenAPI spec version: 1.0
 */
import type { Matricula } from './matricula';

export interface Calificacion {
  /** @nullable */
  descripcion?: string | null;
  id?: number;
  matricula?: Matricula;
  matriculaId?: number;
  nota?: number;
  porcentaje?: number;
}